refactor(search-plugin/tests): extract shared MockKernel + poison_handle to tests/common
The kernel test doubles were copy-pasted into ~10 integration test
files — real drift cost: add_stream and free_buf both had to be sed'd
across every copy. Extract them once into tests/common/mod.rs:

- MockKernel + handle_for (in-memory VFS behind the C-ABI callbacks) —
  shared by index/semantic/title_arm/batch_query/contextual_chunking.
- poison_handle (all-(-1) vtable for tests that must never touch the
  kernel) — shared by query/macro_expand{,_window_cap}/peer_fanout/
  query_expansion/contextual_chunking.

Each file's Harness (embedder / peer-registry / gRPC-server wiring)
stays local — that's genuinely test-specific. index_visibility_e2e is
LEFT as-is on purpose: its kernel is a behaviorally-different stub
(readdir->[], stat->-404), so folding it in would change what it tests.

Net -1489/+355; behaviour-preserving (314 tests pass, unchanged).
