+++ rust/services/search-plugin/tests/contextual_chunking_e2e.rs
@@ -12,13 +12,9 @@
 //! - `chunk.embed_input` (what the embedder sees) is PREFIXED with
 //!   the mock's context, exactly as `apply_contexts` claims.
 
-use std::collections::HashMap;
-use std::ffi::{c_void, CStr};
-use std::os::raw::c_char;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
-use nexus_plugin_abi::KernelHandle;
 use nexus_search_plugin::contextual_chunker::ContextGenerator;
 use nexus_search_plugin::embedder::{EmbedError, Embedder};
 use nexus_search_plugin::index_manager::IndexManager;
@@ -31,76 +27,8 @@ use nexus_search_plugin::service::SearchServiceImpl;
 use tempfile::TempDir;
 use tonic::Request;
 
-// ── Poison KernelHandle — index_documents never calls sys_* ─────
-
-unsafe extern "C" fn poison_read(
-    _: *const c_void,
-    _: *const c_char,
-    _: *mut *mut u8,
-    _: *mut usize,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_write(
-    _: *const c_void,
-    _: *const c_char,
-    _: *const u8,
-    _: usize,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_stat(
-    _: *const c_void,
-    _: *const c_char,
-    _: *mut *mut u8,
-    _: *mut usize,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_readdir(
-    _: *const c_void,
-    _: *const c_char,
-    _: *mut *mut u8,
-    _: *mut usize,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_unlink(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_mkdir(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_rmdir(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_rename(_: *const c_void, _: *const c_char, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn poison_stat_batch(
-    _: *const c_void,
-    _: *const c_char,
-    _: *mut *mut u8,
-    _: *mut usize,
-) -> i32 {
-    -1
-}
-
-fn poison_handle() -> KernelHandle {
-    KernelHandle {
-        sys_read: poison_read,
-        sys_write: poison_write,
-        sys_stat: poison_stat,
-        sys_readdir: poison_readdir,
-        sys_unlink: poison_unlink,
-        sys_mkdir: poison_mkdir,
-        sys_rmdir: poison_rmdir,
-        sys_rename: poison_rename,
-        sys_stat_batch: poison_stat_batch,
-        free_buf: nexus_plugin_abi::nexus_free,
-        kernel_ptr: std::ptr::null(),
-    }
-}
+mod common;
+use common::{handle_for, poison_handle, MockKernel};
 
 // ── Mock embedder that captures embed_input strings ─────────────
 //
@@ -435,192 +363,7 @@ async fn empty_document_never_calls_generator() {
 // site.  Testing only index_documents leaves both walker paths
 // uncovered — a regression that broke apply_contexts inside
 // index_one would ship green.
-//
-// The MockKernel below is the same shape as index_e2e.rs's fixture
-// (adapted here to keep this test file self-contained; a future PR
-// can extract to tests/common/ if a fourth caller shows up).
 
-struct MockFile {
-    bytes: Vec<u8>,
-    mtime_ms: i64,
-}
-
-pub struct MockKernel {
-    files: HashMap<String, MockFile>,
-    dirs: HashMap<String, Vec<(String, u8)>>,
-}
-
-impl MockKernel {
-    fn new() -> Self {
-        Self {
-            files: HashMap::new(),
-            dirs: HashMap::new(),
-        }
-    }
-
-    fn add_file(&mut self, path: &str, bytes: &[u8], mtime_ms: i64) {
-        self.files.insert(
-            path.to_string(),
-            MockFile {
-                bytes: bytes.to_vec(),
-                mtime_ms,
-            },
-        );
-        let (parent, name) = split_parent(path);
-        self.dirs
-            .entry(parent)
-            .or_default()
-            .push((name, 0 /* DT_REG */));
-    }
-}
-
-fn split_parent(path: &str) -> (String, String) {
-    match path.rsplit_once('/') {
-        Some(("", name)) => ("/".to_string(), name.to_string()),
-        Some((parent, name)) => (parent.to_string(), name.to_string()),
-        None => ("/".to_string(), path.to_string()),
-    }
-}
-
-fn into_out_buf(mut v: Vec<u8>, out_buf: *mut *mut u8, out_len: *mut usize) {
-    v.shrink_to_fit();
-    let len = v.len();
-    let ptr = v.as_mut_ptr();
-    std::mem::forget(v);
-    unsafe {
-        *out_buf = ptr;
-        *out_len = len;
-    }
-}
-
-unsafe extern "C" fn mock_sys_read(
-    k: *const c_void,
-    path: *const c_char,
-    out_buf: *mut *mut u8,
-    out_len: *mut usize,
-) -> i32 {
-    let kernel = &*(k as *const MockKernel);
-    let p = match CStr::from_ptr(path).to_str() {
-        Ok(s) => s,
-        Err(_) => return -2,
-    };
-    match kernel.files.get(p) {
-        Some(f) => {
-            into_out_buf(f.bytes.clone(), out_buf, out_len);
-            0
-        }
-        None => -404,
-    }
-}
-
-unsafe extern "C" fn mock_sys_readdir(
-    k: *const c_void,
-    parent: *const c_char,
-    out_json: *mut *mut u8,
-    out_len: *mut usize,
-) -> i32 {
-    let kernel = &*(k as *const MockKernel);
-    let p = match CStr::from_ptr(parent).to_str() {
-        Ok(s) => s,
-        Err(_) => return -2,
-    };
-    match kernel.dirs.get(p) {
-        Some(entries) => {
-            let payload: Vec<serde_json::Value> = entries
-                .iter()
-                .map(|(n, et)| serde_json::json!({ "name": n, "entry_type": et }))
-                .collect();
-            let json = serde_json::to_vec(&payload).expect("mock readdir json");
-            into_out_buf(json, out_json, out_len);
-            0
-        }
-        None => -404,
-    }
-}
-
-unsafe extern "C" fn mock_sys_stat(
-    k: *const c_void,
-    path: *const c_char,
-    out_json: *mut *mut u8,
-    out_len: *mut usize,
-) -> i32 {
-    let kernel = &*(k as *const MockKernel);
-    let p = match CStr::from_ptr(path).to_str() {
-        Ok(s) => s,
-        Err(_) => return -2,
-    };
-    if let Some(f) = kernel.files.get(p) {
-        let payload = serde_json::json!({
-            "path": p,
-            "entry_type": 0,
-            "size": f.bytes.len(),
-            "zone_id": "root",
-            "modified_at_ms": f.mtime_ms,
-        });
-        into_out_buf(serde_json::to_vec(&payload).unwrap(), out_json, out_len);
-        0
-    } else if kernel.dirs.contains_key(p) {
-        let payload = serde_json::json!({
-            "path": p,
-            "entry_type": 1,
-            "size": 0,
-            "zone_id": "root",
-            "modified_at_ms": null,
-        });
-        into_out_buf(serde_json::to_vec(&payload).unwrap(), out_json, out_len);
-        0
-    } else {
-        -404
-    }
-}
-
-unsafe extern "C" fn unused_write(
-    _: *const c_void,
-    _: *const c_char,
-    _: *const u8,
-    _: usize,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_unlink(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_mkdir(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_rmdir(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_rename(_: *const c_void, _: *const c_char, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_stat_batch(
-    _: *const c_void,
-    _: *const c_char,
-    _: *mut *mut u8,
-    _: *mut usize,
-) -> i32 {
-    -1
-}
-
-fn handle_for(k: *const MockKernel) -> KernelHandle {
-    KernelHandle {
-        sys_read: mock_sys_read,
-        sys_write: unused_write,
-        sys_stat: mock_sys_stat,
-        sys_readdir: mock_sys_readdir,
-        sys_unlink: unused_unlink,
-        sys_mkdir: unused_mkdir,
-        sys_rmdir: unused_rmdir,
-        sys_rename: unused_rename,
-        sys_stat_batch: unused_stat_batch,
-        free_buf: nexus_plugin_abi::nexus_free,
-        kernel_ptr: k as *const c_void,
-    }
-}
-
-/// Harness that wraps the mock kernel + service so both index and
-/// refresh tests share one setup shape.
 struct WalkerHarness {
     _dir: TempDir,
     mock: *mut MockKernel,
