+++ rust/services/search-plugin/tests/index_e2e.rs
@@ -17,12 +17,8 @@
 //! `sys_readdir` returns `[{"name","entry_type"}, ...]` and
 //! `sys_stat` returns `{path, entry_type, size, zone_id, modified_at_ms}`.
 
-use std::collections::HashMap;
-use std::ffi::{c_void, CStr};
-use std::os::raw::c_char;
 use std::sync::Arc;
 
-use nexus_plugin_abi::KernelHandle;
 use nexus_search_plugin::index_manager::IndexManager;
 use nexus_search_plugin::search_proto::search_service_server::SearchService;
 use nexus_search_plugin::search_proto::{IndexRequest, QueryRequest, QueryType};
@@ -30,226 +26,8 @@ use nexus_search_plugin::service::SearchServiceImpl;
 use tempfile::TempDir;
 use tonic::Request;
 
-// ── Mock kernel ─────────────────────────────────────────────────
-//
-// The plugin's kernel_io wrappers use these extern "C" callbacks to
-// enumerate + read files.  The mock is a Box'd MockKernel whose raw
-// pointer lives in `KernelHandle.kernel_ptr`; every callback casts
-// the pointer back and answers from the in-memory maps.
-//
-// Memory contract: buffers returned via `out_buf` must be `Vec`s
-// with `cap == len` because the plugin frees via `nexus_free`, which
-// does `Vec::from_raw_parts(ptr, len, len)`.  We `shrink_to_fit()`
-// before `mem::forget` to match.
-
-/// Entry type: 0 = regular file, 1 = directory.  Same constants the
-/// kernel_io module uses (DT_REG / DT_DIR).
-struct FileEntry {
-    bytes: Vec<u8>,
-    mtime_ms: i64,
-}
-
-pub struct MockKernel {
-    // path → file (regular files only; dirs are implicit in `dirs`).
-    files: HashMap<String, FileEntry>,
-    // parent dir → children (name + entry_type).
-    dirs: HashMap<String, Vec<(String, u8)>>,
-}
-
-impl MockKernel {
-    fn new() -> Self {
-        Self {
-            files: HashMap::new(),
-            dirs: HashMap::new(),
-        }
-    }
-
-    fn add_file(&mut self, path: &str, bytes: &[u8], mtime_ms: i64) {
-        self.files.insert(
-            path.to_string(),
-            FileEntry {
-                bytes: bytes.to_vec(),
-                mtime_ms,
-            },
-        );
-        // Splice into parent dir listing.
-        let (parent, name) = split_parent(path);
-        self.dirs
-            .entry(parent)
-            .or_default()
-            .push((name, 0 /* DT_REG */));
-    }
-
-    fn add_dir(&mut self, path: &str) {
-        if !self.dirs.contains_key(path) {
-            self.dirs.insert(path.to_string(), Vec::new());
-        }
-        if path == "/" {
-            return;
-        }
-        let (parent, name) = split_parent(path);
-        // Idempotent: don't double-splice if the dir was already
-        // registered via a nested add_file call.
-        let entries = self.dirs.entry(parent).or_default();
-        if !entries.iter().any(|(n, _)| n == &name) {
-            entries.push((name, 1 /* DT_DIR */));
-        }
-    }
-}
-
-/// Split `/a/b/c.md` into (`/a/b`, `c.md`).  `/foo` yields (`/`, `foo`).
-fn split_parent(path: &str) -> (String, String) {
-    match path.rsplit_once('/') {
-        Some(("", name)) => ("/".to_string(), name.to_string()),
-        Some((parent, name)) => (parent.to_string(), name.to_string()),
-        None => ("/".to_string(), path.to_string()),
-    }
-}
-
-fn into_out_buf(mut v: Vec<u8>, out_buf: *mut *mut u8, out_len: *mut usize) {
-    v.shrink_to_fit();
-    let len = v.len();
-    let ptr = v.as_mut_ptr();
-    std::mem::forget(v);
-    unsafe {
-        *out_buf = ptr;
-        *out_len = len;
-    }
-}
-
-unsafe extern "C" fn mock_sys_read(
-    k: *const c_void,
-    path: *const c_char,
-    out_buf: *mut *mut u8,
-    out_len: *mut usize,
-) -> i32 {
-    let kernel = &*(k as *const MockKernel);
-    let p = match CStr::from_ptr(path).to_str() {
-        Ok(s) => s,
-        Err(_) => return -2,
-    };
-    match kernel.files.get(p) {
-        Some(entry) => {
-            into_out_buf(entry.bytes.clone(), out_buf, out_len);
-            0
-        }
-        None => -404, // NotFound sentinel per plugin-ABI convention
-    }
-}
-
-unsafe extern "C" fn mock_sys_readdir(
-    k: *const c_void,
-    parent_path: *const c_char,
-    out_json: *mut *mut u8,
-    out_len: *mut usize,
-) -> i32 {
-    let kernel = &*(k as *const MockKernel);
-    let p = match CStr::from_ptr(parent_path).to_str() {
-        Ok(s) => s,
-        Err(_) => return -2,
-    };
-    match kernel.dirs.get(p) {
-        Some(entries) => {
-            // JSON encoded per the plugin ABI: [{"name":str,"entry_type":u8}].
-            let payload: Vec<serde_json::Value> = entries
-                .iter()
-                .map(|(name, et)| serde_json::json!({ "name": name, "entry_type": et }))
-                .collect();
-            let json = serde_json::to_vec(&payload).expect("mock readdir json");
-            into_out_buf(json, out_json, out_len);
-            0
-        }
-        None => -404,
-    }
-}
-
-unsafe extern "C" fn mock_sys_stat(
-    k: *const c_void,
-    path: *const c_char,
-    out_json: *mut *mut u8,
-    out_len: *mut usize,
-) -> i32 {
-    let kernel = &*(k as *const MockKernel);
-    let p = match CStr::from_ptr(path).to_str() {
-        Ok(s) => s,
-        Err(_) => return -2,
-    };
-    if let Some(entry) = kernel.files.get(p) {
-        let payload = serde_json::json!({
-            "path": p,
-            "entry_type": 0,
-            "size": entry.bytes.len(),
-            "zone_id": "root",
-            "modified_at_ms": entry.mtime_ms,
-        });
-        into_out_buf(serde_json::to_vec(&payload).unwrap(), out_json, out_len);
-        0
-    } else if kernel.dirs.contains_key(p) {
-        let payload = serde_json::json!({
-            "path": p,
-            "entry_type": 1,
-            "size": 0,
-            "zone_id": "root",
-            "modified_at_ms": null,
-        });
-        into_out_buf(serde_json::to_vec(&payload).unwrap(), out_json, out_len);
-        0
-    } else {
-        -404
-    }
-}
-
-// The write-side callbacks are unused by Index but the plugin ABI
-// requires the full KernelHandle table.  Same poison shape as
-// query_e2e.rs — return -1 so any accidental use fails loudly.
-unsafe extern "C" fn unused_sys_write(
-    _: *const c_void,
-    _: *const c_char,
-    _: *const u8,
-    _: usize,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_sys_unlink(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_sys_mkdir(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_sys_rmdir(_: *const c_void, _: *const c_char) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_sys_rename(
-    _: *const c_void,
-    _: *const c_char,
-    _: *const c_char,
-) -> i32 {
-    -1
-}
-unsafe extern "C" fn unused_sys_stat_batch(
-    _: *const c_void,
-    _: *const c_char,
-    _: *mut *mut u8,
-    _: *mut usize,
-) -> i32 {
-    -1
-}
-
-fn handle_for(kernel: *const MockKernel) -> KernelHandle {
-    KernelHandle {
-        sys_read: mock_sys_read,
-        sys_write: unused_sys_write,
-        sys_stat: mock_sys_stat,
-        sys_readdir: mock_sys_readdir,
-        sys_unlink: unused_sys_unlink,
-        sys_mkdir: unused_sys_mkdir,
-        sys_rmdir: unused_sys_rmdir,
-        sys_rename: unused_sys_rename,
-        sys_stat_batch: unused_sys_stat_batch,
-        free_buf: nexus_plugin_abi::nexus_free,
-        kernel_ptr: kernel as *const c_void,
-    }
-}
+mod common;
+use common::{handle_for, MockKernel};
 
 // ── Harness ─────────────────────────────────────────────────────
